+++ tests/protocols/gmpv208/entities/reports/test_delete_report.py
@@ -17,39 +17,19 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from gvm.errors import GvmError
-from gvm.protocols.gmpv208.entities.report_formats import (
-    ReportFormatType,
-    get_report_format_id_from_string,
-)
 
 
-class GmpDeleteReportFormatTestCase:
+class GmpDeleteReportTestMixin:
     def test_delete(self):
-        self.gmp.delete_report_format('a1')
+        self.gmp.delete_report('a1')
 
         self.connection.send.has_been_called_with(
-            '<delete_report_format report_format_id="a1" ultimate="0"/>'
-        )
-
-    def test_delete_ultimate(self):
-        self.gmp.delete_report_format('a1', ultimate=True)
-
-        self.connection.send.has_been_called_with(
-            '<delete_report_format report_format_id="a1" ultimate="1"/>'
+            '<delete_report report_id="a1"/>'
         )
 
     def test_missing_id(self):
         with self.assertRaises(GvmError):
-            self.gmp.delete_report_format(None)
+            self.gmp.delete_report(None)
 
         with self.assertRaises(GvmError):
-            self.gmp.delete_report_format('')
-
-    def test_delete_with_type(self):
-        self.gmp.delete_report_format(ReportFormatType.SVG)
-
-        report_format_id = get_report_format_id_from_string('svg').value
-        self.connection.send.has_been_called_with(
-            '<delete_report_format '
-            'report_format_id="{}" ultimate="0"/>'.format(report_format_id)
-        )
+            self.gmp.delete_report('')